Code restruction
